% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToVorteks.R
\name{ExportDataframeToVorteks}
\alias{ExportDataframeToVorteks}
\title{Export dataframe to 'Vorteks'}
\usage{
ExportDataframeToVorteks(df, extension, filename = "dataframe_export")
}
\arguments{
\item{df}{dataframe variable}

\item{extension}{must be one of supported extensions (.vvedf, .vpedf, .vdmdf)}

\item{filename}{name of exported file variable}
}
\value{
(No return value, called to perform export process)
}
\description{
'ExportDataframeToVorteks()' will write a file with specified extension
to the current working directory.
In windows, the file will be executed in an attempt to launch 'Vorteks' software.
}
